+++ examples/c/CMakeLists.txt
@@ -53,24 +53,20 @@ function(prepare_executable executable)
   endforeach()
 endfunction()
 
+if(USE_CXX)
+  set(EXAMPLES_SOURCE_DIR ${CMAKE_SOURCE_DIR}/src/cpp)
+else()
+  set(EXAMPLES_SOURCE_DIR ${CMAKE_SOURCE_DIR}/src/c)
+endif()
+
 if(PHI3)
-  if(USE_CXX)
-    add_executable(phi3 ${CMAKE_SOURCE_DIR}/src/cpp/phi3.cpp)
-    prepare_executable(phi3)
-  else()
-    add_executable(phi3 ${CMAKE_SOURCE_DIR}/src/c/phi3.cpp)
-    prepare_executable(phi3)
-  endif()
+  add_executable(phi3 ${EXAMPLES_SOURCE_DIR}/phi3.cpp)
+  prepare_executable(phi3)
 endif()
 
 if(PHI3V)
-  if(USE_CXX)
-    add_executable(phi3v ${CMAKE_SOURCE_DIR}/src/cpp/phi3v.cpp)
-    prepare_executable(phi3v)
-  else()
-    add_executable(phi3v ${CMAKE_SOURCE_DIR}/src/c/phi3v.cpp)
-    prepare_executable(phi3v)
-  endif()
+  add_executable(phi3v ${EXAMPLES_SOURCE_DIR}/phi3v.cpp)
+  prepare_executable(phi3v)
 endif()
 
 if(WHISPER)
