+++ examples/c/src/c/phi3v.cpp
@@ -5,21 +5,8 @@
 #include <cstring>
 #include <fstream>
 #include <memory>
-
+#include "../common.h"
 #include "ort_genai.h"
-
-bool FileExists(const char* path) {
-  return static_cast<bool>(std::ifstream(path));
-}
-
-std::string trim(const std::string& str) {
-  const size_t first = str.find_first_not_of(' ');
-  if (std::string::npos == first) {
-    return str;
-  }
-  const size_t last = str.find_last_not_of(' ');
-  return str.substr(first, (last - first + 1));
-}
 
 // C API Example
 
@@ -126,12 +113,6 @@ void C_API(const char* model_path, const char* execution_provider) {
   OgaDestroyTokenizerStream(tokenizer_stream);
   OgaDestroyMultiModalProcessor(processor);
   OgaDestroyModel(model);
-}
-
-static void print_usage(int /*argc*/, char** argv) {
-  std::cerr << "usage: " << argv[0] << std::endl;
-  std::cerr << "model_path = " << argv[1] << std::endl;
-  std::cerr << "execution_provider = " << argv[2] << std::endl;
 }
 
 int main(int argc, char** argv) {
